% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{add_pull_request}
\alias{add_pull_request}
\title{Title}
\usage{
add_pull_request(
  pull_requests = NULL,
  developer,
  kludge,
  time_to_develop,
  review_status,
  meta_review_status
)
}
\arguments{
\item{pull_requests}{}

\item{developer}{}

\item{kludge}{}

\item{time_to_develop}{}

\item{review_status}{}

\item{meta_review_status}{}
}
\value{

}
\description{
Title
}
