% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{add_task}
\alias{add_task}
\title{Title}
\usage{
add_task(task_backlog, task_type, time_from_now, now)
}
\arguments{
\item{task_backlog}{}

\item{task_type}{}

\item{time_from_now}{}

\item{now}{}
}
\value{

}
\description{
Title
}
