% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{analyse_alerts_and_categories}
\alias{analyse_alerts_and_categories}
\title{Analyses alerts}
\usage{
analyse_alerts_and_categories(
  dir_outside_read = NULL,
  package_start_outside_read = NULL,
  dir_read_all_alerts = NULL,
  pattern_major_read_all_alerts = NULL,
  pattern_minor_read_all_alerts = NULL,
  package_start_read_all_alerts = NULL,
  prefix = NULL
)
}
\value{
categorised alerts
}
\description{
Analyses alerts
}
