% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satd_pmd.R
\name{bind_outputs_correlate_satd}
\alias{bind_outputs_correlate_satd}
\title{Title}
\usage{
bind_outputs_correlate_satd(
  files = c("temp/flink_bags_dist_stem_1.rds", "temp/flink_bags_dist_stem_2.rds",
    "temp/flink_bags_dist_stem_3.rds", "temp/flink_bags_dist_stem_4.rds"),
  output_file = "temp/flink_bags_dist_stem.rds"
)
}
\arguments{
\item{files}{}

\item{output_file}{}
}
\value{

}
\description{
Title
}
