% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_features}
\alias{calculate_features}
\title{calculate the features from the combination of the alerts in the new and the old version}
\usage{
calculate_features(graph_old, graph_new, coordinates)
}
\arguments{
\item{graph_old}{graph related to the old version}

\item{graph_new}{graph related to the new version}

\item{coordinates}{line map}
}
\value{
calculated features
}
\description{
it's called from calculate_features_from_versions
}
