% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_features_from_versions_and_extract_categorised_alerts}
\alias{calculate_features_from_versions_and_extract_categorised_alerts}
\title{Title}
\usage{
calculate_features_from_versions_and_extract_categorised_alerts(
  code_file_new,
  code_file_old,
  pmd_path
)
}
\arguments{
\item{pmd_path}{path to pmd}

\item{dir_old}{old version}

\item{dir_new}{new version}
}
\value{

}
\description{
Title
}
