% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satd_pmd.R
\name{correlate_satd_pmd}
\alias{correlate_satd_pmd}
\title{Title}
\usage{
correlate_satd_pmd(
  code_location = "C:/doutorado/ant/ant-rel-1.10.10",
  output_location = here::here("temp/alerts"),
  pmd_location = here::here("tests/testthat/pmd/bin"),
  alertrules_location = "rulesets/java/quickstart.xml",
 
    blockrules_location = here::here("tests/testthat/data/blockrules/blockrules_simple.xml"),
  output_file,
  bags = FALSE
)
}
\arguments{
\item{code_location}{}

\item{output_location}{}

\item{pmd_location}{}

\item{alertrules_location}{}

\item{blockrules_location}{}
}
\value{

}
\description{
Title
}
