% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{create_pull_request}
\alias{create_pull_request}
\title{Title}
\usage{
create_pull_request(
  developer,
  kludge,
  time_to_develop,
  review_status,
  meta_review_status
)
}
\arguments{
\item{developer}{}

\item{kludge}{}

\item{time_to_develop}{}

\item{review_status}{}

\item{meta_review_status}{}
}
\value{

}
\description{
Title
}
