% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{create_task}
\alias{create_task}
\title{Title}
\usage{
create_task(task_type, earliest_time, player)
}
\arguments{
\item{task_type}{}

\item{earliest_time}{}

\item{player}{}
}
\value{

}
\description{
Title
}
