% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cross_versions}
\alias{cross_versions}
\title{Adds the line map to the dataframe used inside calculate_features_from_versions}
\usage{
cross_versions(examples_executed, diff_content = NA)
}
\arguments{
\item{examples_executed}{dataframe with the versions and their pmd output}
}
\value{
dataframe with line maps info
}
\description{
Adds the line map to the dataframe used inside calculate_features_from_versions
}
