% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{decide_heurist_if_same_alert}
\alias{decide_heurist_if_same_alert}
\title{Decide, with an heuristic, based on features, if a combination of alerts are "the same"}
\usage{
decide_heurist_if_same_alert(clean_calculated_features)
}
\arguments{
\item{clean_calculated_features}{dataframe with one line for each combination of alert.
It may come from calculate_features_from_versions() \%>\% extract_clean_features_from_calculated_features()}
}
\value{
a dataframe with one column "same_alert" with the decision if the two alerts of the combination are the same
}
\description{
given a number of combination of alerts with their features, decides if the alerts are related to the same
}
