% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{decorate_code}
\alias{decorate_code}
\title{Make lines of code nice in a markdown document}
\usage{
decorate_code(strings, size_line_of_code = 80)
}
\arguments{
\item{strings}{lines of code}

\item{size_line_of_code}{number of columns of the lines of code shown}
}
\value{

}
\description{
Make lines of code nice in a markdown document
}
