% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{decorate_code_alerts_mapped}
\alias{decorate_code_alerts_mapped}
\title{Make two versions of lines of code look nice, side by side, in markdown, with their alerts}
\usage{
decorate_code_alerts_mapped(
  strings_old_param,
  alerts_old_param,
  strings_new_param,
  alerts_new_param,
  map_param,
  region_only = FALSE,
  region_size = 3,
  length_alert_name_side_by_side = 14,
  size_line_of_code_side_by_side = 77,
  use_mnemonic = FALSE
)
}
\arguments{
\item{strings_old_param}{source code of the old version}

\item{alerts_old_param}{alerts of the old version}

\item{strings_new_param}{source code of the new version}

\item{alerts_new_param}{alerts of the new version}

\item{map_param}{line map between new and old versions}

\item{region_only}{show only region where alerts are ?}

\item{region_size}{size of the region near alerts to show}

\item{length_alert_name_side_by_side}{length of the alert name in the markdown output}

\item{size_line_of_code_side_by_side}{size of each line of code in the markdown output}
}
\value{
code that can be put on markdown document
}
\description{
Make two versions of lines of code look nice, side by side, in markdown, with their alerts
}
