% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{decorate_code_and_alerts}
\alias{decorate_code_and_alerts}
\title{Make lines of code look nice in markdown, with their alerts}
\usage{
decorate_code_and_alerts(
  strings,
  alerts,
  region_only = FALSE,
  region_size = 3,
  size_line_of_code = 160,
  length_alert_name = 35,
  use_mnemonic = FALSE
)
}
\arguments{
\item{strings}{lines of code}

\item{alerts}{alerts}

\item{region_only}{include only regions near alerts?}

\item{region_size}{size of the region near alerts contained in the output}

\item{size_line_of_code}{size of lines of code in the markdown output}

\item{length_alert_name}{size of lines of alert names in the markdown output}
}
\value{
a string to be used in a markdown file
}
\description{
it's called from read_and_decorate_code
}
