% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_comments_from_code}
\alias{extract_comments_from_code}
\title{Extract comments from java/C source code}
\usage{
extract_comments_from_code(file_path)
}
\arguments{
\item{file_path}{path to the source code file}
}
\value{
dataframe with beginline, endline, begincolumn, endcolumn and comment
}
\description{
Extract comments from java/C source code
}
