% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_diff_pairs_from_diff_file}
\alias{extract_diff_pairs_from_diff_file}
\title{extract from diff files th diff pairs}
\usage{
extract_diff_pairs_from_diff_file(file)
}
\arguments{
\item{file}{diff file}
}
\value{
pair of diff
}
\description{
extract from diff files th diff pairs
}
