% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_piece_of_code}
\alias{extract_piece_of_code}
\title{Extract piece of code}
\usage{
extract_piece_of_code(
  strings_param,
  begin_line,
  end_line,
  begin_column,
  end_column
)
}
\arguments{
\item{strings_param}{code to extract piece from}

\item{begin_line}{line where the piece begins}

\item{end_line}{line where the piece ends}

\item{begin_column}{column where the piece begins}

\item{end_column}{column where the piece ends}
}
\value{
piece of code
}
\description{
Extract piece of code
}
