% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_ast_tree_from_raw_nodes}
\alias{generate_ast_tree_from_raw_nodes}
\title{Generate Abstract Syntax Tree from raw nodes}
\usage{
generate_ast_tree_from_raw_nodes(nodes)
}
\arguments{
\item{nodes}{raw nodes from read_raw_ast_nodes}
}
\value{
the Abstract Syntax Tree, a graph
}
\description{
Generate Abstract Syntax Tree from raw nodes
}
