% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{join_ast_alerts}
\alias{join_ast_alerts}
\title{Join a graph containing the ast and the alerts}
\usage{
join_ast_alerts(ast, alerts)
}
\arguments{
\item{ast}{graph the Abstract Syntax Tree}

\item{alerts}{dataframe containing the alerts}
}
\value{
the graph containing the ast and the related alerts
}
\description{
In this join, the function must consider if there is more than one alert for each node of the ast.
If this is the case, the function must duplicate these nodes
}
