% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_lines}
\alias{map_lines}
\title{Map lines from two files based on the diff result on them}
\usage{
map_lines(file, lines_prev_param, lines_post_param)
}
\arguments{
\item{file}{diff file}

\item{lines_prev_param}{number of lines of left(prev) file}

\item{lines_post_param}{number of lines of right(post) file}
}
\value{

}
\description{
Map lines from two files based on the diff result on them
}
