% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_and_decorate_code}
\alias{read_and_decorate_code}
\title{Reads a code file, decorate and returns a string for a RMarkdown document}
\usage{
read_and_decorate_code(file)
}
\arguments{
\item{file}{file to be read}
}
\value{
a string for a RMarkdown document
}
\description{
Reads a code file, decorate and returns a string for a RMarkdown document
}
