% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_and_decorate_code_and_alerts}
\alias{read_and_decorate_code_and_alerts}
\title{Read a source code and, with its alerts, make it nice to put on a markdown document}
\usage{
read_and_decorate_code_and_alerts(
  file,
  alerts,
  region_only = FALSE,
  region_size = 3,
  use_mnemonic = FALSE
)
}
\arguments{
\item{file}{source code file which matches with the alerts}

\item{alerts}{alerts from an output of calculate_features, output$versions_executed$pmd_output[[i]]}

\item{region_only}{must the output contain only the region with alerts?}

\item{region_size}{how many lines must the region near the alert contained in the output have?}
}
\value{
a nice markdown content to be put on a markdown document
}
\description{
Read a source code and, with its alerts, make it nice to put on a markdown document
}
