% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_and_decorate_code_and_alerts_mapped}
\alias{read_and_decorate_code_and_alerts_mapped}
\title{Reads an old version ans a new version of a source code and returns a string for a RMarkdown document compating them}
\usage{
read_and_decorate_code_and_alerts_mapped(
  file_old,
  alerts_old,
  file_new,
  alerts_new,
  map,
  region_only = FALSE,
  region_size = 3
)
}
\arguments{
\item{file_old}{file containing the old version of a source code}

\item{alerts_old}{alerts for the old version of a source code}

\item{file_new}{file containing the new version of a source code}

\item{alerts_new}{alerts for the new version of a source code}

\item{map}{line map between new and old versions}

\item{region_only}{show only the region near the alerts?}

\item{region_size}{size of the region contained in the output}
}
\value{
string ready for a RMarkdown document
}
\description{
Reads an old version ans a new version of a source code and returns a string for a RMarkdown document compating them
}
