% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_number_of_lines}
\alias{read_number_of_lines}
\title{Return number of lines of the file in the path}
\usage{
read_number_of_lines(dir)
}
\arguments{
\item{dir}{directory, string}
}
\value{
number of lines
}
\description{
Return number of lines of the file in the path
}
