% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_pmd_xml}
\alias{read_pmd_xml}
\title{Read and format XML file generated by PMD}
\usage{
read_pmd_xml(file, include_file_in_output = FALSE)
}
\arguments{
\item{file}{file to be read}
}
\value{
alerts from PMD file
}
\description{
Read and format XML file generated by PMD
}
