% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_raw_ast_nodes}
\alias{read_raw_ast_nodes}
\title{Raw abstract syntax tree from a source code file}
\usage{
read_raw_ast_nodes(
  code_location,
  output_location,
  pmd_location,
  blockrules_location,
  alerts = NA
)
}
\arguments{
\item{code_location}{location of a file containing the source code}

\item{output_location}{temporary output where xml from PMD will be written}

\item{pmd_location}{location of PMD.bat}

\item{blockrules_location}{location of xml containing the blockrules}
}
\value{
a dataframe containing the raw AST, with all the nodes captured from source code
}
\description{
Reads the source code file, runs PMD Source Analyzer and returns the war Abstract Syntax Tree
}
