% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_alerts}
\alias{save_alerts}
\title{Save alerts from a directory with versions}
\usage{
save_alerts(
  dir = here::here("ArgoUML"),
  pmd_path = "pmd/bin/pmd.bat",
  pattern_versions = "^0_[0-9]*"
)
}
\arguments{
\item{dir}{}

\item{pmd_path}{}

\item{pattern_versions}{}
}
\value{

}
\description{
Save alerts from a directory with versions
}
