% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{set_devs_actions}
\alias{set_devs_actions}
\title{Title}
\usage{
set_devs_actions(d1 = "Diligent/Accurate", d2 = "Kludgy/Inaccurate")
}
\arguments{
\item{d1}{}

\item{d2}{}
}
\value{

}
\description{
Title
}
