% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{show_ast}
\alias{show_ast}
\title{Generate a plot from graph with AST}
\usage{
show_ast(
  graph_dfs_tree,
  size_label = 5.5,
  alpha_label = 1,
  node_text_field = "id_alert",
  name_field = "name",
  show_label = TRUE,
  title = "",
  aspect = 1.3,
  nudge_x = 0,
  nudge_y = 0
)
}
\arguments{
\item{graph_dfs_tree}{a graph contained in the output from calculate_features_from_versions}

\item{size_label}{size of the label attached to the node}

\item{alpha_label}{opacity of the label attached to the node}

\item{node_text_field}{name of a column in the output from calculate_features_from_versions used in the label attached to the node}

\item{name_field}{name of a column in the output from calculate_features_from_versions used inside the node}

\item{show_label}{are the labels shown?}

\item{title}{title of the plot}
}
\value{
a plot
}
\description{
Generate a plot from graph with AST
}
