% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{show_latex_raw_ast_nodes}
\alias{show_latex_raw_ast_nodes}
\title{Show nodes as latex}
\usage{
show_latex_raw_ast_nodes(nodes)
}
\arguments{
\item{nodes}{nodes to read}
}
\value{
latex with the nodes
}
\description{
Show nodes as latex
}
