% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare_versions}
\alias{compare_versions}
\title{Compare two versions of a source-code in terms of kludges}
\usage{
compare_versions(dir_old, dir_new)
}
\arguments{
\item{dir_old}{old version}

\item{dir_new}{new version}
}
\value{
alerts categorised in "new", "fixed" and "open"
}
\description{
Lists all the java files, gets the PMD alerts and categorise the alerts in "open", "fixed" and "new"
}
