% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_version_comparisons_comment}
\alias{create_version_comparisons_comment}
\title{Create verstion comparison of comments}
\usage{
create_version_comparisons_comment()
}
\value{
comparison of comments
}
\description{
Create verstion comparison of comments
}
