% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_number_of_lines}
\alias{read_number_of_lines}
\title{Return number of lines of the file}
\usage{
read_number_of_lines(file)
}
\arguments{
\item{file}{file to be read}
}
\value{
number of lines
}
\description{
Return number of lines of the file
}
