% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{report_features}
\alias{report_features}
\title{Reports the features in a latex table}
\usage{
report_features(features_df, caption)
}
\arguments{
\item{features_df}{features info from calculate_features_from_versions}

\item{caption}{caption of the table}
}
\value{
a table in latex
}
\description{
Reports the features in a latex table
}
